Remove nested classes
