+++ src/seedu/addressbook/data/person/Block.java
@@ -1,0 +1,13 @@
+package seedu.addressbook.data.person;
+
+public class Block {
+	private String value;
+	
+	public Block(String block){
+		value = block;
+	}
+	
+	public String getValue(){
+		return value;
+	}
+}
