+++ src/seedu/addressbook/data/person/PostalCode.java
@@ -1,0 +1,13 @@
+package seedu.addressbook.data.person;
+
+class PostalCode{
+	private String value;
+	
+	public PostalCode(String postalCode){
+		value = postalCode;
+	}
+	
+	public String getValue(){
+		return value;
+	}
+}
