+++ src/seedu/addressbook/data/person/Street.java
@@ -1,0 +1,14 @@
+package seedu.addressbook.data.person;
+
+public class Street{
+	private String value;
+	
+	public Street(String street){
+		value = street;
+	}
+	
+	public String getValue(){
+		return value;
+	}
+}
+
