+++ src/seedu/addressbook/data/person/Unit.java
@@ -1,0 +1,13 @@
+package seedu.addressbook.data.person;
+
+class Unit{
+	private String value;
+	
+	public Unit(String unit){
+		value = unit;
+	}
+	
+	public String getValue(){
+		return value;
+	}
+}
